Move action methods to concern
